+++ zcash_client_backend/CHANGELOG.md
@@ -28,6 +28,13 @@ workspace.
   - `LightdInfo` has added fields `upgrade_name`, `upgrade_height`, and `lighwallet_protocol_version`
   - `GetMempoolTxRequest` (previously named `Exclude`) has added field `pool_types`
 - `zcash_client_backend::fees::MetaSource`
+- `zcash_client_backend::scanning`:
+  - `full` module, providing full-block scanning.
+  - `Nullifiers::unspent`
+  - `Nullifiers::update_with`
+- `zcash_client_backend::sync`:
+  - `decryptor` module, behind the `sync-decryptor` feature flag, providing a
+    Tokio-based batch decryption engine for full blocks and transactions.
 - `zcash_client_backend::wallet`:
   - `transparent` module, behind the `transparent-inputs` feature flag.
   - `Note::receiver`